requirements.txt
absl-py==0.9.0
astor==0.8.1
boto3==1.12.49
botocore==1.15.49
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dataclasses==0.7
docutils==0.15.2
filelock==3.0.12
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
jmespath==0.9.5
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
pkg-resources==0.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.41
scipy==1.4.1
sentencepiece==0.1.86
six==1.14.0
tensorboard==2.1.1
tensorflow-gpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tokenizers==0.5.2
tqdm==4.45.0
transformers==2.8.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
